Serialize witness (#6)
* add witness store and fetch

* remove halo2_gadgets

* store pk related and witness related data

* only store/fetch mapping in assembly

* save witness in a quick manner

* using u32 for Assembly for better serialization

* add named advices

* imporve pk loading speed

* disable randomize for named columns

* update rust toolchain

* fix: assert selector number equals zero

+++ halo2_proofs/src/dev.rs
@@ -1032,7 +1032,7 @@ impl<F: FieldExt> MockProver<F> {
                     // value is preserved by the mapping.
                     values.iter().enumerate().filter_map(move |(row, cell)| {
                         let original_cell = original(column, row);
-                        let permuted_cell = original(cell.0, cell.1);
+                        let permuted_cell = original(cell.0 as usize, cell.1 as usize);
                         if original_cell == permuted_cell {
                             None
                         } else {
